+++ src/main/java/com/atlassian/braid/source/MapGraphQLError.java
@@ -5,10 +5,16 @@ import graphql.ErrorType;
 import graphql.GraphQLError;
 import graphql.language.SourceLocation;
 
+import java.util.Arrays;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Collectors;
+
+import static java.util.Collections.unmodifiableSet;
+import static java.util.stream.Collectors.toMap;
 
 /**
  * Turns a map representing the data of a GraphQL error into a real GraphQLError
@@ -21,6 +27,7 @@ public class MapGraphQLError implements GraphQLError {
     private final List<SourceLocation> locations;
     private final ErrorType errorType;
     private final List<Object> path;
+    private final Map<String, Object> extensions;
 
     public MapGraphQLError(Map<String, Object> error) {
         this.message = Optional.ofNullable(error.get("message")).map(String.class::cast).orElse("Unknown error");
@@ -34,6 +41,10 @@ public class MapGraphQLError implements GraphQLError {
                 .orElse(null);
         this.errorType = ErrorType.DataFetchingException;
         this.path = Optional.ofNullable(error.get("path")).map(BraidObjects::<List<Object>>cast).orElse(null);
+        this.extensions = Optional.ofNullable(error.get("extensions"))
+                .map(BraidObjects::<Map<String, Object>>cast)
+                .orElse(null);
+
     }
 
     @Override
@@ -54,5 +65,10 @@ public class MapGraphQLError implements GraphQLError {
     @Override
     public List<Object> getPath() {
         return path;
+    }
+
+    @Override
+    public Map<String, Object> getExtensions() {
+        return extensions;
     }
 }